diff, restructure
