+++ dictknife/diff/__init__.py
@@ -1,0 +1,2 @@
+from .rows import diff_rows  # noqa
+from .diff import diff  # noqa
